requirements.txt
ollama==0.6.1
discord==2.3.2
python-dotenv==1.2.1
numpy==2.4.2
google-generativeai
